Assignment 1 || DS & Algo 2
Section 1
Define a data structure using C#, which represents a department in a big organization. The department has the following properties:
Name
Manager name
List of employees
List of sub-departments
The list of employees must contain names or objects of a class defined to represent employees. The list of departments must contain instances of the same data structure.

Use the data structure to create data representing a big company with at least 15 departments and at least 4 levels of sub-departments. The top structure will represent the whole company. Create methods for adding and removing sub-departments and employees to departments. Implement methods for moving employees from one department to another and for moving sub-departments. Develop functionality for calculating the number of employees in a department, including the sub-departments. 

Create a user-friendly interface to enter and display the data.